+++ api.php
@@ -291,6 +291,22 @@ if($_REQUEST['key'] && $_REQUEST['do']){
 	
 	//CALLBACKS
 	if($_REQUEST['do'] == 'callback'){ 
+	//SKIP MATERIAL FROM MAKE FORMULA
+		if($_REQUEST['action'] == 'skipMaterial' && $_REQUEST['fid'] &&  $_REQUEST['id']){
+			$fid = mysqli_real_escape_string($conn, $_REQUEST['fid']);
+			$id = mysqli_real_escape_string($conn, $_REQUEST['id']);
+			$ingID = mysqli_real_escape_string($conn, $_REQUEST['ingId']);
+			$notes = mysqli_real_escape_string($conn, $_REQUEST['notes']) ?: "-";
+		
+			if(mysqli_query($conn, "UPDATE makeFormula SET skip = '1', notes = '$notes' WHERE fid = '$fid' AND id = '$id'")){
+				$response['success'] = $_POST['ing'].' skipped from the formulation';
+			} else {
+				$response['error'] = 'Error skipping the ingredient';
+			}
+			
+			echo json_encode($response);
+			return;
+		}
 		if( $_REQUEST['action'] == 'makeFormula'){
 	
 			$fid = mysqli_real_escape_string($conn, $_REQUEST['fid']);
@@ -311,7 +327,14 @@ if($_REQUEST['key'] && $_REQUEST['do']){
 				echo json_encode($response);
 				return;
 			}
-								 
+			
+			if((double)$_REQUEST['q'] == 0.00){
+				$response['success'] = false;
+				$response['message'] = 'Please add quantity';
+				echo json_encode($response);
+				return;
+			}
+							 
 			$q = trim($_REQUEST['q']);
 			$notes = mysqli_real_escape_string($conn, $_REQUEST['notes']);