+++ db/pvault.sql
@@ -71,6 +71,7 @@ CREATE TABLE `makeFormula` (
  `overdose` double(8,4) NOT NULL DEFAULT 0.0000,
  `originalQuantity` double(8,4) DEFAULT NULL,
  `notes` MEDIUMTEXT NOT NULL,
+ `skip` INT NOT NULL DEFAULT '0', 
  `toAdd` int(11) NOT NULL,
  PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;