+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Increase update interval check from PV Scale to 5s
 - Added separate notes field for Formula Make
 - When a formula is marked as completed it generates a document in formula attachments
+- Added API key to the data sent to PV Scale
+- Add material skip when making a formula
 
 ### Version 9.8
 - Added phpMyAdmin definition in the docker compose