+++ releasenotes.md
@@ -1,5 +1,6 @@
 Whats New in v9.9
 --------------------------
+- Skip a mterial when you making a formula
 - Increase update interval check from PV Scale to 5s
 - Added separate notes field for Formula Make
 - When a formula is marked as completed it generates a document in formula attachments